[Feat/#304] :wrench: Refactor: PlayerScore 외부 클래스로 이동
