+++ chatty-be/src/main/java/com/chatty/chatty/game/service/GameService.java
@@ -12,9 +12,9 @@ import com.chatty.chatty.game.controller.dto.model.MarkRequest.AnswerDTO;
 import com.chatty.chatty.game.controller.dto.model.MarkResponse;
 import com.chatty.chatty.game.domain.AnswerData;
 import com.chatty.chatty.game.domain.AnswerData.PlayerAnswerData;
+import com.chatty.chatty.game.domain.PlayerScore;
 import com.chatty.chatty.game.domain.QuizData;
 import com.chatty.chatty.game.domain.ScoreData;
-import com.chatty.chatty.game.domain.ScoreData.PlayerScore;
 import com.chatty.chatty.game.domain.SubmitStatus;
 import com.chatty.chatty.game.domain.UsersSubmitStatus;
 import com.chatty.chatty.game.repository.AnswerRepository;
